pyproject.toml
[project]
name = "tg-tracking-habits"
version = "0.1.0"
description = ""
authors = [
    {name = "Илья Тарасов",email = "ilyatarasov2006@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pytelegrambotapi (>=4.26.0,<5.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aiohttp (>=3.11.18,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
